+++ src/lib.rs
@@ -170,7 +170,7 @@ The *fixed* crate requires rustc version 1.44.0 or later.
 
 ## Optional features
 
-The *fixed* crate has four optional feature:
+The *fixed* crate has these optional feature:
 
  1. `az`, disabled by default. This implements the cast traits
     provided by the [*az* crate].
@@ -182,6 +182,12 @@ The *fixed* crate has four optional feature:
  4. `std`, disabled by default. This is for features that are not
     possible under `no_std`: currently the implementation of the
     [`Error`] trait for [`ParseFixedError`].
+ 5. `serde-str`, disabled by default. Fixed-point numbers are
+    serialized as strings showing the value when using human-readable
+    formats. This feature requires the `serde` and the `std` optional
+    features. **Warning:** numbers serialized when this feature is
+    enabled cannot be deserialized when this feature is disabled, and
+    vice versa.
 
 To enable features, you can add the dependency like this to
 [*Cargo.toml*]:
@@ -210,13 +216,6 @@ There are three experimental feature:
     assertions are enabled. (The plan is to always enable this
     functionality but remove the experimental feature in version
     1.5.0.)
- 3. `serde-str`, disabled by default. Fixed-point numbers are
-    serialized as strings showing the value when using human-readable
-    formats. This feature requires the `serde` and the `std` optional
-    features. (The plan is to upgrade this to an optional feature in
-    version 1.5.0.) **Warning:** numbers serialized when this feature
-    is enabled cannot be deserialized when this feature is disabled,
-    and vice versa.
 
 ## License